requirements.txt
dj-database-url==2.0.0
Django==4.2.3
mysqlclient==2.2.4
python-decouple==3.8
isort==5.13.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1